+++ packages/google_maps_flutter/google_maps_flutter/example/lib/map_click.dart
@@ -34,6 +34,7 @@ class _MapClickBodyState extends State<_MapClickBody> {
   GoogleMapController? mapController;
   LatLng? _lastTap;
   LatLng? _lastLongPress;
+  PointOfInterestId? _lastPointOfInterestTap;
 
   @override
   Widget build(BuildContext context) {
@@ -48,6 +49,11 @@ class _MapClickBodyState extends State<_MapClickBody> {
       onLongPress: (LatLng pos) {
         setState(() {
           _lastLongPress = pos;
+        });
+      },
+      onPointOfInterestTap: (PointOfInterestId placeId) {
+        setState(() {
+          _lastPointOfInterestTap = placeId;
         });
       },
     );
@@ -68,6 +74,12 @@ class _MapClickBodyState extends State<_MapClickBody> {
     if (mapController != null) {
       final String lastTap = 'Tap:\n${_lastTap ?? ""}\n';
       final String lastLongPress = 'Long press:\n${_lastLongPress ?? ""}';
+      final String lastPointOfInterestTap =
+          'Point of interest tap:\n${_lastPointOfInterestTap?.value}';
+
+      if(_lastPointOfInterestTap != null) {
+      }
+
       columnChildren.add(Center(
           child: Text(
         lastTap,
@@ -86,6 +98,16 @@ class _MapClickBodyState extends State<_MapClickBody> {
       columnChildren.add(Center(
           child: Text(
         _lastLongPress != null ? 'Long pressed' : '',
+        textAlign: TextAlign.center,
+      )));
+      columnChildren.add(Center(
+          child: Text(
+        lastPointOfInterestTap,
+        textAlign: TextAlign.center,
+      )));
+      columnChildren.add(Center(
+          child: Text(
+        _lastPointOfInterestTap != null ? 'Point of interest tapped' : '',
         textAlign: TextAlign.center,
       )));
     }