+++ packages/google_maps_flutter/google_maps_flutter/lib/src/controller.dart
@@ -75,6 +75,9 @@ class GoogleMapController {
         .onCircleTap(mapId: mapId)
         .listen((CircleTapEvent e) => _googleMapState.onCircleTap(e.value));
     GoogleMapsFlutterPlatform.instance
+        .onPointOfInterestTap(mapId: mapId)
+        .listen((MapPointOfInterestTapEvent e) => _googleMapState.onPointOfInterestTap(e.value));
+    GoogleMapsFlutterPlatform.instance
         .onTap(mapId: mapId)
         .listen((MapTapEvent e) => _googleMapState.onTap(e.position));
     GoogleMapsFlutterPlatform.instance.onLongPress(mapId: mapId).listen(