+++ packages/google_maps_flutter/google_maps_flutter/lib/src/google_map.dart
@@ -127,6 +127,7 @@ class GoogleMap extends StatefulWidget {
     this.onTap,
     this.onLongPress,
     this.cloudMapId,
+    this.onPointOfInterestTap,
   });
 
   /// Callback method for when the map is ready to be used.
@@ -243,6 +244,9 @@ class GoogleMap extends StatefulWidget {
 
   /// Called every time a [GoogleMap] is long pressed.
   final ArgumentCallback<LatLng>? onLongPress;
+
+  /// Called every time a [PointOfInterest] is clicked.
+  final ArgumentCallback<PointOfInterestId>? onPointOfInterestTap;
 
   /// True if a "My Location" layer should be shown on the map.
   ///
@@ -541,6 +545,14 @@ class _GoogleMapState extends State<GoogleMap> {
     final ArgumentCallback<LatLng>? onLongPress = widget.onLongPress;
     if (onLongPress != null) {
       onLongPress(position);
+    }
+  }
+
+  void onPointOfInterestTap(PointOfInterestId placeId) {
+    final ArgumentCallback<PointOfInterestId>? onPointOfInterestTap =
+        widget.onPointOfInterestTap;
+    if (onPointOfInterestTap != null) {
+      onPointOfInterestTap(placeId);
     }
   }
 }