+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapController.java
@@ -35,6 +35,7 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.MapStyleOptions;
 import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.PointOfInterest;
 import com.google.android.gms.maps.model.Polygon;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.maps.android.clustering.ClusterManager;
@@ -592,6 +593,12 @@ class GoogleMapController
     methodChannel.invokeMethod("map#onLongPress", arguments);
   }
 
+  public void onPoiClick(PointOfInterest poi) {
+    final Map<String, Object> arguments = new HashMap<>(2);
+    arguments.put("placeId", poi.placeId);
+    methodChannel.invokeMethod("map#onPointOfInterestTap", arguments);
+  }
+
   @Override
   public void onCameraMoveStarted(int reason) {
     final Map<String, Object> arguments = new HashMap<>(2);
@@ -687,6 +694,7 @@ class GoogleMapController
     googleMap.setOnCircleClickListener(listener);
     googleMap.setOnMapClickListener(listener);
     googleMap.setOnMapLongClickListener(listener);
+    googleMap.setOnPoiClickListener(listener);
   }
 
   @VisibleForTesting