+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/GoogleMapController.m
@@ -580,6 +580,16 @@
          arguments:@{@"position" : [FLTGoogleMapJSONConversions arrayFromLocation:coordinate]}];
 }
 
+- (void)mapView:(GMSMapView *)mapView
+    didTapPOIWithPlaceID:(NSString *)placeID
+                    name:(NSString *)name
+                location:(CLLocationCoordinate2D)location {
+  [self.channel invokeMethod:@"map#onPointOfInterestTap"
+                   arguments:@{
+                     @"placeId" : placeID,
+                   }];
+}
+
 - (void)interpretMapOptions:(NSDictionary *)data {
   NSArray *cameraTargetBounds = data[@"cameraTargetBounds"];
   if (cameraTargetBounds && cameraTargetBounds != (id)[NSNull null]) {