+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.9.0
+
+* Adds point of interest tap event handler.
+
 ## 2.8.0
 
 * Adds support for marker clustering.