+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.7.0
+
+* Adds point of interest tap event handler.
+
 ## 2.6.0
 
 * Updates the minimum allowed verison of the Google Maps SDK to 8.4, for privacy